+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Multi Translator
 
+## 2.28.4 - 2026-07-23
+
+### Fixed
+
+- fix: ensure existingSettings are array before merging (fixes #130)
+
 ## 2.28.3 - 2026-06-25
 
 ### Fixed